voxcore/voxquery/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & SQL Generation
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.1.0
sqlalchemy==2.0.23
sqlglot==23.0.0
sqlparse==0.4.4
urllib3==1.26.18

# Data Warehouses
snowflake-sqlalchemy==1.5.1
snowflake-connector-python>=3.0.0
psycopg2-binary==2.9.9
pyodbc==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.3
polars==0.19.12

# Security
python-dotenv==1.0.0
cryptography==41.0.7
pyjwt>=2.8.0
python-jose==3.3.0
passlib==1.7.4
itsdangerous>=2.0.0

# Observability & Logging
langsmith==0.0.83
structlog==23.2.0
sentry-sdk==1.39.1

# Results Formatting
openpyxl>=3.0.0
plotly==5.18.0
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
pylint==3.0.3
mypy==1.7.1
ruff==0.1.11

# Utils
python-multipart==0.0.6
aiofiles==23.2.1
click==8.1.7
